requirements.txt
Pillow
requests
flask
gallery-dl
urllib3
numpy
imagehash
